DSRC Demo Session
An example of using DNA Sequence Reads Compression (DSRC) for compressing FASTQ files to a smaller size. Website for the tool can be found in the asciicast. NOTE: you will need to tweak the dsrc Makefile to compile the binary on Centos6 (I'll create an RPM for it via the Halocaridina repo in mid-May).